generic-orchestration-samples/setup/setup_register_es/requirements.txt
cloudshell-orch-core
cloudshell-sandbox-reporter